full restructured
