--- pygreppy/pygreppymain.py
@@ -1,15 +1,0 @@
-import sys
-from pygreppy import Args, parse, usage
-
-
-def main():
-    args = Args(sys.argv[1:])
-    if args.args == 1:
-        usage()
-        sys.exit(1)
-    else:
-        print('\n' + parse(args))
-
-
-if __name__ == '__main__':
-    main()
